Vet clinic ca - upload
Me and Joshua decided to split out our partnership for this CA due our available time to work not matching, since both of us are still working, and our timetables are very different. Therefore, we were not getting time to seat together and discuss the strategies we would like to take for this project. At this stage we could work together but from here on we will be working individually. Each one of us will submit a independent folder.